add: ドラフト記事の認可制御

+++ frontend/src/pages/admin/AdminDraftIndexArticle.tsx
@@ -35,7 +35,9 @@ export const AdminDraftIndexArticle: React.FC = () => {
 
   const executeDeleteArticle = async () => {
     if (selectedArticle) {
-      await deleteArticleHooks.deleteArticle(selectedArticle.id)
+      if (currentUser?.role == ADMIN) {
+        await deleteArticleHooks.deleteArticle(selectedArticle.id)
+      }
       await fetchArticlesHooks.fetchArticles(true)
       setIsOpenSnackbar(true)
     } else {
